added a hook to the fragment in the main activity
