+++ .idea/libraries/design_24_1_0.xml
@@ -1,0 +1,12 @@
+<component name="libraryTable">
+  <library name="design-24.1.0">
+    <CLASSES>
+      <root url="jar://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.android.support/design/24.1.0/jars/classes.jar!/" />
+      <root url="file://$PROJECT_DIR$/app/build/intermediates/exploded-aar/com.android.support/design/24.1.0/res" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/Library/Android/sdk/extras/android/m2repository/com/android/support/design/24.1.0/design-24.1.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
